fixed: compile warning
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@30333 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/addons/Repository.cpp
@@ -102,7 +102,7 @@ CStdString CRepository::Checksum()
   CStdString checksum;
   try
   {
-    char* temp = new char[file.GetLength()+1];
+    char* temp = new char[(size_t)file.GetLength()+1];
     file.Read(temp,file.GetLength());
     temp[file.GetLength()] = 0;
     checksum = temp;
